refactor(Logger): Add code improvements and change debug logger folder
