+++ ui/src/components/debug_logger/script.js
