+++ ui/src/components/debug_logger/style.scss
