+++ ui/src/components/mod.rs
@@ -1,5 +1,6 @@
 pub mod calldialog;
 pub mod chat;
+pub mod debug_logger;
 pub mod friends;
 pub mod media;
 pub mod settings;
