+++ ui/src/components/settings/sub_pages/audio.rs
@@ -2,7 +2,7 @@ use dioxus::prelude::*;
 use kit::elements::switch::Switch;
 use shared::language::get_local_text;
 
-use crate::{components::settings::SettingSection, logger::logger::Logger};
+use crate::{components::settings::SettingSection, logger::Logger};
 
 #[allow(non_snake_case)]
 pub fn AudioSettings(cx: Scope) -> Element {
