+++ ui/src/components/settings/sub_pages/privacy.rs
@@ -5,7 +5,7 @@ use kit::{
 };
 use shared::language::get_local_text;
 
-use crate::{components::settings::SettingSection, logger::logger::Logger};
+use crate::{components::settings::SettingSection, logger::Logger};
 
 #[allow(non_snake_case)]
 pub fn PrivacySettings(cx: Scope) -> Element {
