+++ ui/src/window_manager/mod.rs
@@ -8,7 +8,7 @@ use tokio::sync::{
 };
 
 use crate::{
-    logger::logger::Logger,
+    logger::Logger,
     state::{Action, State},
 };
